refactor: refactored part of the codebase to improve structure and reduce circular dependencies
