+++ src/handlers/__init__.py
@@ -1,0 +1,13 @@
+from .authorization import authorized_only
+from .birthdays import send_birthdays_month, back_to_birthdays
+from .commands import send_main_menu, proceed_authorize_users, toggle_admin_mode, temp_authorize_user, \
+    send_mass_message,  remind_password, proceed_mass_message, old_button_handler
+# from .departments import ...
+from .employees import add_employee, proceed_add_employee_data, skip_phone, skip_email, skip_username, skip_dob, \
+    send_profile, edit_employee
+from .google_forms import callback_ans, cancel_form_filling
+from .links import send_business_process, send_business_processes_menu, add_link, proceed_add_link_data, send_form, \
+    send_helpdesk, show_helpdesk_password, edit_link, proceed_edit_link_data, delete_link_confirmation, delete_link, \
+    back_to_send_links
+from .main_menu import send_knowledge_base, send_business_processes, send_birthdays, send_contacts_menu, thanks_menu, \
+    send_useful_links, send_form
