+++ src/integrations/__init__.py
@@ -1,0 +1,3 @@
+from .crm_api_functions import get_employee_pass_from_crm, add_employee_to_crm, delete_employee_from_crm, \
+    update_employee_in_crm
+from .telethon_functions import proceed_find_user_id, send_photo, decrypt_session, remove_user_from_chat
